paper2019/requirements.txt
torchaudio
visdom
wget
librosa
torch-stft

